+++ src/components/Markdown/index.js
@@ -1,0 +1,7 @@
+import ReactMarkdown from 'react-markdown'
+
+import React from 'react'
+
+export default ({children}) => (
+  <ReactMarkdown source={children} />
+)
