+++ src/screens/Instructors/components/InstructorsList/components/InstructorTeaser/index.js
