+++ src/screens/Instructors/components/InstructorsList/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import {map, size} from 'lodash'
 import {Heading} from 'egghead-ui'
 import {noInstructorsDescriptionText} from 'utils/text'
-import InstructorSummary from './components/InstructorSummary'
+import InstructorTeaser from './components/InstructorTeaser'
 
 export default ({title, instructors}) => (
   <div>
@@ -18,7 +18,7 @@ export default ({title, instructors}) => (
                   key={instructor.id}
                   className='bb b--navy pa3'
                 >
-                  <InstructorSummary instructor={instructor} />
+                  <InstructorTeaser instructor={instructor} />
                 </div>
               ))}
             </div>
